+++ app/src/main/java/com/bintangpoetra/sumbanginaja/presentation/splash/SplashFragment.kt
@@ -21,7 +21,7 @@ class SplashFragment : BaseFragment<FragmentSplashBinding>() {
             if (pref.isLogin) {
                 findNavController().navigate(R.id.action_splashFragment_to_homeFragment)
             } else {
-                findNavController().navigate(R.id.action_splashFragment_to_loginFragment)
+                findNavController().navigate(R.id.action_splashFragment_to_onboardingFragment)
             }
         }, 1500)
     }
