+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -1,6 +1,52 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="@dimen/dimen_0dp"
+        android:layout_height="@dimen/dimen_0dp"
+        android:scaleType="centerCrop"
+        android:src="@drawable/onboardingbg"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btnStart"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/action_start"
+        android:textAllCaps="false"
+        android:layout_marginBottom="@dimen/dimen_32dp"
+        android:layout_marginHorizontal="@dimen/dimen_32dp"
+        app:backgroundTint="@color/white"
+        app:layout_constraintBottom_toBottomOf="@+id/imageView"
+        style="@style/SumbanginAjaButton.BlackText"/>
+
+    <TextView
+        android:id="@+id/txWelcoming"
+        android:text="@string/label_welcoming_onboarding"
+        android:textColor="@color/white"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="@dimen/dimen_32dp"
+        android:layout_marginHorizontal="@dimen/dimen_32dp"
+        app:layout_constraintBottom_toTopOf="@+id/btnStart"
+        style="@style/SumbanginAjaTextStyle.Regular.Heading3"/>
+
+    <TextView
+        android:id="@+id/txSumbanginAja"
+        android:text="@string/app_name"
+        android:textColor="@color/white"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="@dimen/dimen_16dp"
+        android:layout_marginHorizontal="@dimen/dimen_32dp"
+        app:layout_constraintBottom_toTopOf="@+id/txWelcoming"
+        style="@style/SumbanginAjaTextStyle.Bold.Heading1"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
