+++ app/src/main/res/values/button_styles.xml
@@ -14,6 +14,10 @@
         <item name="android:textAppearance">@style/SumbanginAjaTextStyle.Regular.Heading3.White</item>
     </style>
 
+    <style name="SumbanginAjaButton.BlackText">
+        <item name="android:textAppearance">@style/SumbanginAjaTextStyle.Regular.Heading3</item>
+    </style>
+
     <style name="SumbanginAjaButtonOutline" parent="Widget.Material3.Button.OutlinedButton">
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">wrap_content</item>
