+++ app/src/main/res/values/dimens.xml
@@ -10,4 +10,5 @@
     <dimen name="dimen_12dp">12dp</dimen>
     <dimen name="dimen_300dp">300dp</dimen>
     <dimen name="dimen_36dp">36dp</dimen>
+    <dimen name="dimen_0dp">0dp</dimen>
 </resources>
